Made create/resolve ticket tests more stable
Removed the checks for the number of open/resolved tickets. The number
does not always update immediately and during that time somebody else
could change the number. A "bulk edit" will not immediately happen and
may happen during the test.

Instead just check if the ticket is now in the list of open/closed
tickets, and for resolved tickets check the state as well.

The "ticket bins" can be changed by users, so their link names are not
stable. Instead use the links in the "Tickets" menu.

For closed tickets, ensure that there is an open ticket as a given. If
there isn't one, then tell the user to run the create ticket scenario
first and try again later.

Made some finders work better.