Added test for CSV export validation
Set up a Selenium profile and use a helper module based on
http://collectiveidea.com/blog/archives/2012/01/27/testing-file-downloads-with-capybara-and-chromedriver/
so that downloads happen without a manual "open with" dialog, are saved
to a correct location and wait until it is finished.

After that it was pretty easy to set up. Not sure if it works in all
situations though, so needs more testing.