Moved a feature, split scenarios, fixed some steps
Renamed 'create ticket' to 'add a new ticket' (the title of the page
after you click create new ticket) so that it is the first scenario that
is run. This makes it more likely that the other features pass too.

Split the 'change profile picture' into two scenarios so that it is
shown that they work independently.

Fixed steps for 'log in' which still had arguments when the regex did
not have anymore.