Use all tickets, and fixed number of tickets
Allow some scenarios to run on any ticket rather than open ticket. This
makes it more likely to pass because it can find a ticket to use on.

After 30 tickets the text in the 'search sentence' changes due to which
another b element appears, which we don't want. Therefore use the last
element for the number.